+++ app/api/v1/routers/root.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, HTTPException
 from app.constants.schemas import get_thought_schema
-from app.services.openai_service import get_structured_response
+from app.services.openai import get_structured_response
 
 router = APIRouter(tags=["root"]) # no prefix; mounted at "/"
