+++ app/services/prompting.py
@@ -1,6 +1,5 @@
 from datetime import datetime
 import json
-import random
 import textwrap
 from typing import Any, List, Mapping, Optional, Sequence
