+++ app/services/utility.py
@@ -1,9 +1,8 @@
 import asyncio
 import os
 import random
-from motor.motor_asyncio import AsyncIOMotorClient
-from app.constants.constants import AGENT_COLLECTION, AGENT_LITE_COLLECTION, AGENT_NAME_PROPERTY, MIN_EMOTION_VALUE, EMOTIONAL_DECAY_RATE
-from app.services.data_service import get_database, grab_self
+from app.constants.constants import AGENT_NAME_PROPERTY, MIN_EMOTION_VALUE, EMOTIONAL_DECAY_RATE
+from app.services.database import grab_self, update_agent_emotions
 
 async def emotion_decay_loop(decay_rate: int, lite_mode: bool):
     '''
@@ -12,7 +11,6 @@ async def emotion_decay_loop(decay_rate: int, lite_mode: bool):
         :param decay_rate: The number of seconds between each loop of the function
         :param lite_mode: Whether we are dealing with lite_mode emotions
     '''
-    db = get_database()
     while True:
         try:
             self = await grab_self(os.getenv('BOT_NAME'), lite_mode)
@@ -33,14 +31,13 @@ async def emotion_decay_loop(decay_rate: int, lite_mode: bool):
                     self["emotional_status"]["emotions"][emotion]["value"] = data["value"]
 
             if (lite_mode):
-                db[AGENT_LITE_COLLECTION].update_one({"name": self[AGENT_NAME_PROPERTY]}, { "$set": {"emotional_status": self["emotional_status"] }})
+                await update_agent_emotions(self[AGENT_NAME_PROPERTY], self["emotional_status"])
             else: 
-                db[AGENT_COLLECTION].update_one({"name": self[AGENT_NAME_PROPERTY]}, { "$set": {"emotional_status": self["emotional_status"] }})
+                await update_agent_emotions(self[AGENT_NAME_PROPERTY], self["emotional_status"], False)
 
         except Exception as e:
             print(f"Error - Emotional decay: {e}")
 
-        # Wait for the specified decay rate
         await asyncio.sleep(decay_rate)
 
 async def start_emotion_decay():
